Re-do duplivert mesh randomly
The duplivert mesh that I had created for the bottles was too regular
(made it from a decimate of the sea mesh), so instead, I made a fresh
copy of the sea mesh and deleted vertices by random selection until the
density looked about right. Then I had to recreate faces to get proper
normals on the vertices, so I just did a fill and triangulated the
result.